Data.Vector.Hybrid

Description

A hybrid Vector lets you make a Vector which is 'partially unboxed', by making a Vector out of two other Vector types and using each for its corresponding side of a Vector of pairs.

This enables you to work with a mixture of boxed and unboxed data when you go to use something like vector-algorithms

Documentation

data Vector

data MVector

Accessors

Length information

length :: Vector u a => Vector u v (a, b) -> Int

O(1) Yield the length of the vector.

null :: Vector u a => Vector u v (a, b) -> Bool

O(1) Test whether a vector if empty

Indexing

(!) :: (Vector u a, Vector v b) => Vector u v (a, b) -> Int -> (a, b)

O(1) Indexing

(!?) :: (Vector u a, Vector v b) => Vector u v (a, b) -> Int -> Maybe (a, b)

O(1) Safe indexing

head :: (Vector u a, Vector v b) => Vector u v (a, b) -> (a, b)

O(1) First element

last :: (Vector u a, Vector v b) => Vector u v (a, b) -> (a, b)

O(1) Last element

unsafeIndex :: (Vector u a, Vector v b) => Vector u v (a, b) -> Int -> (a, b)

O(1) Unsafe indexing without bounds checking

unsafeHead :: (Vector u a, Vector v b) => Vector u v (a, b) -> (a, b)

O(1) First element without checking if the vector is empty

unsafeLast :: (Vector u a, Vector v b) => Vector u v (a, b) -> (a, b)

O(1) Last element without checking if the vector is empty

Monadic indexing

indexM :: (Vector u a, Vector v b, Monad m) => Vector u v (a, b) -> Int -> m (a, b)

O(1) Indexing in a monad.

The monad allows operations to be strict in the vector when necessary. Suppose vector copying is implemented like this:

 copy mv v = ... write mv i (v ! i) ...

For lazy vectors, v ! i would not be evaluated which means that mv would unnecessarily retain a reference to v in each element written.

With indexM, copying can be implemented like this instead:

 copy mv v = ... do
                   x <- indexM v i
                   write mv i x

Here, no references to v are retained because indexing (but not the elements) is evaluated eagerly.

headM :: (Vector u a, Vector v b, Monad m) => Vector u v (a, b) -> m (a, b)

O(1) First element of a vector in a monad. See indexM for an explanation of why this is useful.

lastM :: (Vector u a, Vector v b, Monad m) => Vector u v (a, b) -> m (a, b)

O(1) Last element of a vector in a monad. See indexM for an explanation of why this is useful.

unsafeIndexM :: (Vector u a, Vector v b, Monad m) => Vector u v (a, b) -> Int -> m (a, b)

O(1) Indexing in a monad without bounds checks. See indexM for an explanation of why this is useful.

unsafeHeadM :: (Vector u a, Vector v b, Monad m) => Vector u v (a, b) -> m (a, b)

O(1) First element in a monad without checking for empty vectors. See indexM for an explanation of why this is useful.

unsafeLastM :: (Vector u a, Vector v b, Monad m) => Vector u v (a, b) -> m (a, b)

O(1) Last element in a monad without checking for empty vectors. See indexM for an explanation of why this is useful.

Extracting subvectors (slicing)

slice

Arguments

:: (Vector u a, Vector v b)

=> Int
i starting index

-> Int
n length

-> Vector u v (a, b)

-> Vector u v (a, b)

O(1) Yield a slice of the vector without copying it. The vector must contain at least i+n elements.

init :: (Vector u a, Vector v b) => Vector u v (a, b) -> Vector u v (a, b)

O(1) Yield all but the last element without copying. The vector may not be empty.

tail :: (Vector u a, Vector v b) => Vector u v (a, b) -> Vector u v (a, b)

O(1) Yield all but the first element without copying. The vector may not be empty.

take :: (Vector u a, Vector v b) => Int -> Vector u v (a, b) -> Vector u v (a, b)

O(1) Yield at the first n elements without copying. The vector may contain less than n elements in which case it is returned unchanged.

drop :: (Vector u a, Vector v b) => Int -> Vector u v (a, b) -> Vector u v (a, b)

O(1) Yield all but the first n elements without copying. The vector may contain less than n elements in which case an empty vector is returned.

splitAt :: (Vector u a, Vector v b) => Int -> Vector u v (a, b) -> (Vector u v (a, b), Vector u v (a, b))

O(1) Yield the first n elements paired with the remainder without copying.

Note that splitAt n v is equivalent to (take n v, drop n v) but slightly more efficient.

unsafeSlice

Arguments

:: (Vector u a, Vector v b)

=> Int
i starting index

-> Int
n length

-> Vector u v (a, b)

-> Vector u v (a, b)

O(1) Yield a slice of the vector without copying. The vector must contain at least i+n elements but this is not checked.

unsafeInit :: (Vector u a, Vector v b) => Vector u v (a, b) -> Vector u v (a, b)

O(1) Yield all but the last element without copying. The vector may not be empty but this is not checked.

unsafeTail :: (Vector u a, Vector v b) => Vector u v (a, b) -> Vector u v (a, b)

O(1) Yield all but the first element without copying. The vector may not be empty but this is not checked.

unsafeTake :: (Vector u a, Vector v b) => Int -> Vector u v (a, b) -> Vector u v (a, b)

O(1) Yield the first n elements without copying. The vector must contain at least n elements but this is not checked.

unsafeDrop :: (Vector u a, Vector v b) => Int -> Vector u v (a, b) -> Vector u v (a, b)

O(1) Yield all but the first n elements without copying. The vector must contain at least n elements but this is not checked.

Construction

Initialisation

empty :: (Vector u a, Vector v b) => Vector u v (a, b)

O(1) Empty vector

singleton :: (Vector u a, Vector v b) => (a, b) -> Vector u v (a, b)

O(1) Vector with exactly one element

replicate :: (Vector u a, Vector v b) => Int -> (a, b) -> Vector u v (a, b)

O(n) Vector of the given length with the same value in each position

generate :: (Vector u a, Vector v b) => Int -> (Int -> (a, b)) -> Vector u v (a, b)

O(n) Construct a vector of the given length by applying the function to each index

iterateN :: (Vector u a, Vector v b) => Int -> ((a, b) -> (a, b)) -> (a, b) -> Vector u v (a, b)

O(n) Apply function n times to value. Zeroth element is original value.

Monadic initialisation

replicateM :: (Monad m, Vector u a, Vector v b) => Int -> m (a, b) -> m (Vector u v (a, b))

O(n) Execute the monadic action the given number of times and store the results in a vector.

generateM :: (Monad m, Vector u a, Vector v b) => Int -> (Int -> m (a, b)) -> m (Vector u v (a, b))

O(n) Construct a vector of the given length by applying the monadic action to each index

create :: (Vector u a, Vector v b) => (forall s. ST s (Mutable (Vector u v) s (a, b))) -> Vector u v (a, b)

Execute the monadic action and freeze the resulting vector.

 create (do { v <- new 2; write v 0 'a'; write v 1 'b'; return v }) = <a,b>

Unfolding

unfoldr :: (Vector u a, Vector v b) => (c -> Maybe ((a, b), c)) -> c -> Vector u v (a, b)

O(n) Construct a vector by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

 unfoldr (\n -> if n == 0 then Nothing else Just (n,n-1)) 10
  = <10,9,8,7,6,5,4,3,2,1>

unfoldrN :: (Vector u a, Vector v b) => Int -> (c -> Maybe ((a, b), c)) -> c -> Vector u v (a, b)

O(n) Construct a vector with at most n by repeatedly applying the generator function to the a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

 unfoldrN 3 (\n -> Just (n,n-1)) 10 = <10,9,8>

constructN :: (Vector u a, Vector v b) => Int -> (Vector u v (a, b) -> (a, b)) -> Vector u v (a, b)

O(n) Construct a vector with n elements by repeatedly applying the generator function to the already constructed part of the vector.

 constructN 3 f = let a = f <> ; b = f <a> ; c = f <a,b> in f <a,b,c>

constructrN :: (Vector u a, Vector v b) => Int -> (Vector u v (a, b) -> (a, b)) -> Vector u v (a, b)

O(n) Construct a vector with n elements from right to left by repeatedly applying the generator function to the already constructed part of the vector.

 constructrN 3 f = let a = f <> ; b = f<a> ; c = f <b,a> in f <c,b,a>

Concatenation

cons :: (Vector u a, Vector v b) => (a, b) -> Vector u v (a, b) -> Vector u v (a, b)

O(n) Prepend an element

snoc :: (Vector u a, Vector v b) => Vector u v (a, b) -> (a, b) -> Vector u v (a, b)

O(n) Append an element

(++) :: (Vector u a, Vector v b) => Vector u v (a, b) -> Vector u v (a, b) -> Vector u v (a, b)

O(m+n) Concatenate two vectors

concat :: (Vector u a, Vector v b) => [Vector u v (a, b)] -> Vector u v (a, b)

O(n) Concatenate all vectors in the list

Restricting memory usage

force :: (Vector u a, Vector v b) => Vector u v (a, b) -> Vector u v (a, b)

O(n) Yield the argument but force it not to retain any extra memory, possibly by copying it.

This is especially useful when dealing with slices. For example:

 force (slice 0 2 <huge vector>)

Here, the slice retains a reference to the huge vector. Forcing it creates a copy of just the elements that belong to the slice and allows the huge vector to be garbage collected.

Modifying vectors

Bulk updates

(//)

Arguments

:: (Vector u a, Vector v b)

=> Vector u v (a, b)
initial vector (of length m)

-> [(Int, (a, b))]
list of index/value pairs (of length n)

-> Vector u v (a, b)

O(m+n) For each pair (i,a) from the list, replace the vector element at position i by a.

 <5,9,2,7> // [(2,1),(0,3),(2,8)] = <3,9,8,7>

unsafeUpd :: (Vector u a, Vector v b) => Vector u v (a, b) -> [(Int, (a, b))] -> Vector u v (a, b)

Same as (//) but without bounds checking.

Accumulations

accum

Arguments

:: (Vector u a, Vector v b)

=> ((a, b) -> c -> (a, b))
accumulating function f

-> Vector u v (a, b)
initial vector (of length m)

-> [(Int, c)]
list of index/value pairs (of length n)

-> Vector u v (a, b)

O(m+n) For each pair (i,c) from the list, replace the vector element a at position i by f a c.

 accum (+) <5,9,2> [(2,4),(1,6),(0,3),(1,7)] = <5+3, 9+6+7, 2+4>

unsafeAccum :: (Vector u a, Vector v b) => ((a, b) -> c -> (a, b)) -> Vector u v (a, b) -> [(Int, c)] -> Vector u v (a, b)

Same as accum but without bounds checking.

Permutations

reverse :: (Vector u a, Vector v b) => Vector u v (a, b) -> Vector u v (a, b)

O(n) Reverse a vector

Safe destructive updates

modify :: (Vector u a, Vector v b) => (forall s. Mutable (Vector u v) s (a, b) -> ST s ()) -> Vector u v (a, b) -> Vector u v (a, b)

Apply a destructive operation to a vector. The operation will be performed in place if it is safe to do so and will modify a copy of the vector otherwise.

 modify (\v -> write v 0 'x') (replicate 3 'a') = <'x','a','a'>

Elementwise operations

Mapping

map :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d)) -> Vector u v (a, b) -> Vector u v (c, d)

O(n) Map a function over a vector

imap :: (Vector u a, Vector v b, Vector u c, Vector v d) => (Int -> (a, b) -> (c, d)) -> Vector u v (a, b) -> Vector u v (c, d)

O(n) Apply a function to every element of a vector and its index

concatMap :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> Vector u v (c, d)) -> Vector u v (a, b) -> Vector u v (c, d)

Map a function over a vector and concatenate the results.

Monadic mapping

mapM :: (Monad m, Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> m (c, d)) -> Vector u v (a, b) -> m (Vector u v (c, d))

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results

mapM_ :: (Monad m, Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> m (c, d)) -> Vector u v (a, b) -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results

forM :: (Monad m, Vector u a, Vector v b, Vector u c, Vector v d) => Vector u v (a, b) -> ((a, b) -> m (c, d)) -> m (Vector u v (c, d))

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results. Equvalent to flip mapM.

forM_ :: (Monad m, Vector u a, Vector v b, Vector u c, Vector v d) => Vector u v (a, b) -> ((a, b) -> m (c, d)) -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results. Equivalent to flip mapM_.

Zipping

zipWith :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c') => ((a, a') -> (b, b') -> (c, c')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c')

O(min(m,n)) Zip two vectors with the given function.

zipWith3 :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c', Vector u d, Vector v d') => ((a, a') -> (b, b') -> (c, c') -> (d, d')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c') -> Vector u v (d, d')

Zip three vectors with the given function.

zipWith4 :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c', Vector u d, Vector v d', Vector u e, Vector v e') => ((a, a') -> (b, b') -> (c, c') -> (d, d') -> (e, e')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c') -> Vector u v (d, d') -> Vector u v (e, e')

zipWith5 :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c', Vector u d, Vector v d', Vector u e, Vector v e', Vector u f, Vector v f') => ((a, a') -> (b, b') -> (c, c') -> (d, d') -> (e, e') -> (f, f')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c') -> Vector u v (d, d') -> Vector u v (e, e') -> Vector u v (f, f')

zipWith6 :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c', Vector u d, Vector v d', Vector u e, Vector v e', Vector u f, Vector v f', Vector u g, Vector v g') => ((a, a') -> (b, b') -> (c, c') -> (d, d') -> (e, e') -> (f, f') -> (g, g')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c') -> Vector u v (d, d') -> Vector u v (e, e') -> Vector u v (f, f') -> Vector u v (g, g')

izipWith :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c') => (Int -> (a, a') -> (b, b') -> (c, c')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c')

O(min(m,n)) Zip two vectors with a function that also takes the elements' indices.

izipWith3 :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c', Vector u d, Vector v d') => (Int -> (a, a') -> (b, b') -> (c, c') -> (d, d')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c') -> Vector u v (d, d')

Zip three vectors and their indices with the given function.

izipWith4 :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c', Vector u d, Vector v d', Vector u e, Vector v e') => (Int -> (a, a') -> (b, b') -> (c, c') -> (d, d') -> (e, e')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c') -> Vector u v (d, d') -> Vector u v (e, e')

izipWith5 :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c', Vector u d, Vector v d', Vector u e, Vector v e', Vector u f, Vector v f') => (Int -> (a, a') -> (b, b') -> (c, c') -> (d, d') -> (e, e') -> (f, f')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c') -> Vector u v (d, d') -> Vector u v (e, e') -> Vector u v (f, f')

izipWith6 :: (Vector u a, Vector v a', Vector u b, Vector v b', Vector u c, Vector v c', Vector u d, Vector v d', Vector u e, Vector v e', Vector u f, Vector v f', Vector u g, Vector v g') => (Int -> (a, a') -> (b, b') -> (c, c') -> (d, d') -> (e, e') -> (f, f') -> (g, g')) -> Vector u v (a, a') -> Vector u v (b, b') -> Vector u v (c, c') -> Vector u v (d, d') -> Vector u v (e, e') -> Vector u v (f, f') -> Vector u v (g, g')

Monadic zipping

zipWithM :: (Monad m, Vector u a, Vector v b, Vector u c, Vector v d, Vector u e, Vector v f) => ((a, b) -> (c, d) -> m (e, f)) -> Vector u v (a, b) -> Vector u v (c, d) -> m (Vector u v (e, f))

O(min(m,n)) Zip the two vectors with the monadic action and yield a vector of results

zipWithM_ :: (Monad m, Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> m e) -> Vector u v (a, b) -> Vector u v (c, d) -> m ()

O(min(m,n)) Zip the two vectors with the monadic action and ignore the results

Working with predicates

Filtering

filter :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> Vector u v (a, b)

O(n) Drop elements that do not satisfy the predicate

ifilter :: (Vector u a, Vector v b) => (Int -> (a, b) -> Bool) -> Vector u v (a, b) -> Vector u v (a, b)

O(n) Drop elements that do not satisfy the predicate which is applied to values and their indices

filterM :: (Monad m, Vector u a, Vector v b) => ((a, b) -> m Bool) -> Vector u v (a, b) -> m (Vector u v (a, b))

O(n) Drop elements that do not satisfy the monadic predicate

takeWhile :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> Vector u v (a, b)

O(n) Yield the longest prefix of elements satisfying the predicate without copying.

dropWhile :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> Vector u v (a, b)

O(n) Drop the longest prefix of elements that satisfy the predicate without copying.

Partitioning

partition :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> (Vector u v (a, b), Vector u v (a, b))

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The relative order of the elements is preserved at the cost of a sometimes reduced performance compared to unstablePartition.

unstablePartition :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> (Vector u v (a, b), Vector u v (a, b))

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The order of the elements is not preserved but the operation is often faster than partition.

span :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> (Vector u v (a, b), Vector u v (a, b))

O(n) Split the vector into the longest prefix of elements that satisfy the predicate and the rest without copying.

break :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> (Vector u v (a, b), Vector u v (a, b))

O(n) Split the vector into the longest prefix of elements that do not satisfy the predicate and the rest without copying.

Searching

elem :: (Vector u a, Vector v b, Eq a, Eq b) => (a, b) -> Vector u v (a, b) -> Bool

O(n) Check if the vector contains an element

notElem :: (Vector u a, Vector v b, Eq a, Eq b) => (a, b) -> Vector u v (a, b) -> Bool

O(n) Check if the vector does not contain an element (inverse of elem)

find :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> Maybe (a, b)

O(n) Yield Just the first element matching the predicate or Nothing if no such element exists.

findIndex :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> Maybe Int

O(n) Yield Just the index of the first element matching the predicate or Nothing if no such element exists.

elemIndex :: (Vector u a, Vector v b, Eq a, Eq b) => (a, b) -> Vector u v (a, b) -> Maybe Int

O(n) Yield Just the index of the first occurence of the given element or Nothing if the vector does not contain the element. This is a specialised version of findIndex.

Folding

foldl :: (Vector u a, Vector v b) => (r -> (a, b) -> r) -> r -> Vector u v (a, b) -> r

O(n) Left fold

foldl1 :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> (a, b)) -> Vector u v (a, b) -> (a, b)

O(n) Left fold on non-empty vectors

foldl' :: (Vector u a, Vector v b) => (r -> (a, b) -> r) -> r -> Vector u v (a, b) -> r

O(n) Left fold with strict accumulator

foldl1' :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> (a, b)) -> Vector u v (a, b) -> (a, b)

O(n) Left fold on non-empty vectors with strict accumulator

foldr :: (Vector u a, Vector v b) => ((a, b) -> r -> r) -> r -> Vector u v (a, b) -> r

O(n) Right fold

foldr1 :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> (a, b)) -> Vector u v (a, b) -> (a, b)

O(n) Right fold on non-empty vectors

foldr' :: (Vector u a, Vector v b) => ((a, b) -> r -> r) -> r -> Vector u v (a, b) -> r

O(n) Right fold with a strict accumulator

foldr1' :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> (a, b)) -> Vector u v (a, b) -> (a, b)

O(n) Right fold on non-empty vectors with strict accumulator

ifoldl :: (Vector u a, Vector v b) => (r -> Int -> (a, b) -> r) -> r -> Vector u v (a, b) -> r

O(n) Left fold (function applied to each element and its index)

ifoldl' :: (Vector u a, Vector v b) => (r -> Int -> (a, b) -> r) -> r -> Vector u v (a, b) -> r

O(n) Left fold with strict accumulator (function applied to each element and its index)

ifoldr :: (Vector u a, Vector v b) => (Int -> (a, b) -> r -> r) -> r -> Vector u v (a, b) -> r

O(n) Right fold (function applied to each element and its index)

ifoldr' :: (Vector u a, Vector v b) => (Int -> (a, b) -> r -> r) -> r -> Vector u v (a, b) -> r

O(n) Right fold with strict accumulator (function applied to each element and its index)

Specialised folds

all :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> Bool

O(n) Check if all elements satisfy the predicate.

any :: (Vector u a, Vector v b) => ((a, b) -> Bool) -> Vector u v (a, b) -> Bool

O(n) Check if any element satisfies the predicate.

maximum :: (Vector u a, Vector v b, Ord a, Ord b) => Vector u v (a, b) -> (a, b)

O(n) Yield the maximum element of the vector. The vector may not be empty.

maximumBy :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> Ordering) -> Vector u v (a, b) -> (a, b)

O(n) Yield the maximum element of the vector according to the given comparison function. The vector may not be empty.

minimum :: (Vector u a, Vector v b, Ord a, Ord b) => Vector u v (a, b) -> (a, b)

O(n) Yield the minimum element of the vector. The vector may not be empty.

minimumBy :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> Ordering) -> Vector u v (a, b) -> (a, b)

O(n) Yield the minimum element of the vector according to the given comparison function. The vector may not be empty.

minIndex :: (Vector u a, Vector v b, Ord a, Ord b) => Vector u v (a, b) -> Int

O(n) Yield the index of the minimum element of the vector. The vector may not be empty.

minIndexBy :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> Ordering) -> Vector u v (a, b) -> Int

O(n) Yield the index of the minimum element of the vector according to the given comparison function. The vector may not be empty.

maxIndex :: (Vector u a, Vector v b, Ord a, Ord b) => Vector u v (a, b) -> Int

O(n) Yield the index of the maximum element of the vector. The vector may not be empty.

maxIndexBy :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> Ordering) -> Vector u v (a, b) -> Int

O(n) Yield the index of the maximum element of the vector according to the given comparison function. The vector may not be empty.

Monadic folds

foldM :: (Monad m, Vector u a, Vector v b) => (r -> (a, b) -> m r) -> r -> Vector u v (a, b) -> m r

O(n) Monadic fold

foldM' :: (Monad m, Vector u a, Vector v b) => (r -> (a, b) -> m r) -> r -> Vector u v (a, b) -> m r

O(n) Monadic fold with strict accumulator

fold1M :: (Monad m, Vector u a, Vector v b) => ((a, b) -> (a, b) -> m (a, b)) -> Vector u v (a, b) -> m (a, b)

O(n) Monadic fold over non-empty vectors

fold1M' :: (Monad m, Vector u a, Vector v b) => ((a, b) -> (a, b) -> m (a, b)) -> Vector u v (a, b) -> m (a, b)

O(n) Monadic fold over non-empty vectors with strict accumulator

foldM_ :: (Monad m, Vector u a, Vector v b) => (r -> (a, b) -> m r) -> r -> Vector u v (a, b) -> m ()

O(n) Monadic fold that discards the result

foldM'_ :: (Monad m, Vector u a, Vector v b) => (r -> (a, b) -> m r) -> r -> Vector u v (a, b) -> m ()

O(n) Monadic fold with strict accumulator that discards the result

fold1M_ :: (Monad m, Vector u a, Vector v b) => ((a, b) -> (a, b) -> m (a, b)) -> Vector u v (a, b) -> m ()

O(n) Monadic fold over non-empty vectors that discards the result

fold1M'_ :: (Monad m, Vector u a, Vector v b) => ((a, b) -> (a, b) -> m (a, b)) -> Vector u v (a, b) -> m ()

O(n) Monadic fold over non-empty vectors with strict accumulator that discards the result

Prefix sums (scans)

prescanl :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (a, b)) -> (a, b) -> Vector u v (c, d) -> Vector u v (a, b)

O(n) Prescan

 prescanl f z = init . scanl f z

Example: prescanl (+) 0 <1,2,3,4> = <0,1,3,6>

prescanl' :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (a, b)) -> (a, b) -> Vector u v (c, d) -> Vector u v (a, b)

O(n) Prescan with strict accumulator

postscanl :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (a, b)) -> (a, b) -> Vector u v (c, d) -> Vector u v (a, b)

O(n) Scan

 postscanl f z = tail . scanl f z

Example: postscanl (+) 0 <1,2,3,4> = <1,3,6,10>

postscanl' :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (a, b)) -> (a, b) -> Vector u v (c, d) -> Vector u v (a, b)

O(n) Scan with strict accumulator

scanl :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (a, b)) -> (a, b) -> Vector u v (c, d) -> Vector u v (a, b)

O(n) Haskell-style scan

 scanl f z <x1,...,xn> = <y1,...,y(n+1)>
   where y1 = z
         yi = f y(i-1) x(i-1)

Example: scanl (+) 0 <1,2,3,4> = <0,1,3,6,10>

scanl' :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (a, b)) -> (a, b) -> Vector u v (c, d) -> Vector u v (a, b)

O(n) Haskell-style scan with strict accumulator

scanl1 :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> (a, b)) -> Vector u v (a, b) -> Vector u v (a, b)

O(n) Scan over a non-empty vector

 scanl f <x1,...,xn> = <y1,...,yn>
   where y1 = x1
         yi = f y(i-1) xi

scanl1' :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> (a, b)) -> Vector u v (a, b) -> Vector u v (a, b)

O(n) Scan over a non-empty vector with a strict accumulator

prescanr :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (c, d)) -> (c, d) -> Vector u v (a, b) -> Vector u v (c, d)

O(n) Right-to-left prescan

 prescanr f z = reverse . prescanl (flip f) z . reverse

prescanr' :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (c, d)) -> (c, d) -> Vector u v (a, b) -> Vector u v (c, d)

O(n) Right-to-left prescan with strict accumulator

postscanr :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (c, d)) -> (c, d) -> Vector u v (a, b) -> Vector u v (c, d)

O(n) Right-to-left scan

postscanr' :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (c, d)) -> (c, d) -> Vector u v (a, b) -> Vector u v (c, d)

O(n) Right-to-left scan with strict accumulator

scanr :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (c, d)) -> (c, d) -> Vector u v (a, b) -> Vector u v (c, d)

O(n) Right-to-left Haskell-style scan

scanr' :: (Vector u a, Vector v b, Vector u c, Vector v d) => ((a, b) -> (c, d) -> (c, d)) -> (c, d) -> Vector u v (a, b) -> Vector u v (c, d)

O(n) Right-to-left Haskell-style scan with strict accumulator

scanr1 :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> (a, b)) -> Vector u v (a, b) -> Vector u v (a, b)

O(n) Right-to-left scan over a non-empty vector

scanr1' :: (Vector u a, Vector v b) => ((a, b) -> (a, b) -> (a, b)) -> Vector u v (a, b) -> Vector u v (a, b)

O(n) Right-to-left scan over a non-empty vector with a strict accumulator

Conversions

projectFst :: Vector u v (a, b) -> u a

projectSnd :: Vector u v (a, b) -> v b

unsafeZip :: u a -> v b -> Vector u v (a, b)

Warning: The vectors are assumed to have the same length. This is not checked!

Lists

toList :: (Vector u a, Vector v b) => Vector u v (a, b) -> [(a, b)]

O(n) Convert a vector to a list

fromList :: (Vector u a, Vector v b) => [(a, b)] -> Vector u v (a, b)

O(n) Convert a list to a vector

fromListN :: (Vector u a, Vector v b) => Int -> [(a, b)] -> Vector u v (a, b)

O(n) Convert the first n elements of a list to a vector

 fromListN n xs = fromList (take n xs)

Other vector types

convert :: (Vector v a, Vector w a) => v a -> w a

Mutable vectors

freeze :: (Vector u a, Vector v b, PrimMonad m) => Mutable (Vector u v) (PrimState m) (a, b) -> m (Vector u v (a, b))

O(n) Yield an immutable copy of the mutable vector.

thaw :: (Vector u a, Vector v b, PrimMonad m) => Vector u v (a, b) -> m (Mutable (Vector u v) (PrimState m) (a, b))

O(n) Yield a mutable copy of the immutable vector.

copy :: (Vector u a, Vector v b, PrimMonad m) => Mutable (Vector u v) (PrimState m) (a, b) -> Vector u v (a, b) -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length.

unsafeFreeze :: (Vector u a, Vector v b, PrimMonad m) => Mutable (Vector u v) (PrimState m) (a, b) -> m (Vector u v (a, b))

O(1) Unsafe convert a mutable vector to an immutable one without copying. The mutable vector may not be used after this operation.

unsafeThaw :: (Vector u a, Vector v b, PrimMonad m) => Vector u v (a, b) -> m (Mutable (Vector u v) (PrimState m) (a, b))

O(1) Unsafely convert an immutable vector to a mutable one without copying. The immutable vector may not be used after this operation.

unsafeCopy :: (Vector u a, Vector v b, PrimMonad m) => Mutable (Vector u v) (PrimState m) (a, b) -> Vector u v (a, b) -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length. This is not checked.