Data.Vector.Mixed

Description

A mixed Vector lets you make a Vector out of any other vector type you have lying around, and all of the combinators are defined to allow you to freely mix input vector type wherever possible.

This enables you to work with a mixture of boxed and unboxed data.

Mixed vectors

data Vector

data MVector

A MVector s a is mutable vector that could have any vector type underneath

class (Typeable2 mv, Typeable1 v, mv ~ Mutable v, MVector mv a, Vector v a) => Mixed mv v a | mv -> v, v -> mv where

Vector doesn't provide a way to recover the type of the immutable vector from the mutable vector type

This would otherwise prevent us from finishing the implementation of basicUnsafeFreeze in Vector

This class captures the invariants necessary to hide the choice of vector type from the user in such a way that we can go from mutable vector to immutabl vector and back again.

Methods

mmix :: mv s a -> MVector s a

mix :: v a -> Vector a

Accessors

Length information

length :: Vector v a => v a -> Int

O(1) Yield the length of the vector.

null :: Vector v a => v a -> Bool

O(1) Test whether a vector if empty

Indexing

(!) :: Vector v a => v a -> Int -> a

O(1) Indexing

(!?) :: Vector v a => v a -> Int -> Maybe a

O(1) Safe indexing

head :: Vector v a => v a -> a

O(1) First element

last :: Vector v a => v a -> a

O(1) Last element

unsafeIndex :: Vector v a => v a -> Int -> a

O(1) Unsafe indexing without bounds checking

unsafeHead :: Vector v a => v a -> a

O(1) First element without checking if the vector is empty

unsafeLast :: Vector v a => v a -> a

O(1) Last element without checking if the vector is empty

Monadic indexing

indexM :: (Monad m, Vector v a) => v a -> Int -> m a

O(1) Indexing in a monad.

The monad allows operations to be strict in the vector when necessary. Suppose vector copying is implemented like this:

 copy mv v = ... write mv i (v ! i) ...

For lazy vectors, v ! i would not be evaluated which means that mv would unnecessarily retain a reference to v in each element written.

With indexM, copying can be implemented like this instead:

 copy mv v = ... do
                   x <- indexM v i
                   write mv i x

Here, no references to v are retained because indexing (but not the elements) is evaluated eagerly.

headM :: (Monad m, Vector v a) => v a -> m a

O(1) First element of a vector in a monad. See indexM for an explanation of why this is useful.

lastM :: (Monad m, Vector v a) => v a -> m a

O(1) Last element of a vector in a monad. See indexM for an explanation of why this is useful.

unsafeIndexM :: (Monad m, Vector v a) => v a -> Int -> m a

O(1) Indexing in a monad without bounds checks. See indexM for an explanation of why this is useful.

unsafeHeadM :: (Monad m, Vector v a) => v a -> m a

O(1) First element in a monad without checking for empty vectors. See indexM for an explanation of why this is useful.

unsafeLastM :: (Monad m, Vector v a) => v a -> m a

O(1) Last element in a monad without checking for empty vectors. See indexM for an explanation of why this is useful.

Extracting subvectors (slicing)

slice

Arguments

:: Mixed u v a

=> Int
i starting index

-> Int
n length

-> v a

-> Vector a

O(1) Yield a slice of the vector without copying it. The vector must contain at least i+n elements.

init :: Mixed u v a => v a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty.

tail :: Mixed u v a => v a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty.

take :: Mixed u v a => Int -> v a -> Vector a

O(1) Yield at the first n elements without copying. The vector may contain less than n elements in which case it is returned unchanged.

drop :: Mixed u v a => Int -> v a -> Vector a

O(1) Yield all but the first n elements without copying. The vector may contain less than n elements in which case an empty vector is returned.

splitAt :: Mixed u v a => Int -> v a -> (Vector a, Vector a)

O(1) Yield the first n elements paired with the remainder without copying.

Note that splitAt n v is equivalent to (take n v, drop n v) but slightly more efficient.

unsafeSlice

Arguments

:: Mixed u v a

=> Int
i starting index

-> Int
n length

-> v a

-> Vector a

O(1) Yield a slice of the vector without copying. The vector must contain at least i+n elements but this is not checked.

unsafeInit :: Mixed u v a => v a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty but this is not checked.

unsafeTail :: Mixed u v a => v a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty but this is not checked.

unsafeTake :: Mixed u v a => Int -> v a -> Vector a

O(1) Yield the first n elements without copying. The vector must contain at least n elements but this is not checked.

unsafeDrop :: Mixed u v a => Int -> v a -> Vector a

O(1) Yield all but the first n elements without copying. The vector must contain at least n elements but this is not checked.

Construction

Initialisation

empty :: Vector a

O(1) Empty vector

singleton :: a -> Vector a

O(1) Vector with exactly one element

replicate :: Int -> a -> Vector a

O(n) Vector of the given length with the same value in each position

generate :: Int -> (Int -> a) -> Vector a

O(n) Construct a vector of the given length by applying the function to each index

iterateN :: Int -> (a -> a) -> a -> Vector a

O(n) Apply function n times to value. Zeroth element is original value.

Monadic initialisation

replicateM :: Monad m => Int -> m a -> m (Vector a)

O(n) Execute the monadic action the given number of times and store the results in a vector.

generateM :: Monad m => Int -> (Int -> m a) -> m (Vector a)

O(n) Construct a vector of the given length by applying the monadic action to each index

create :: Mixed u v a => (forall s. ST s (u s a)) -> Vector a

Execute the monadic action and freeze the resulting vector.

 create (do { v <- new 2; write v 0 'a'; write v 1 'b'; return v }) = <a,b>

Unfolding

unfoldr :: (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

 unfoldr (\n -> if n == 0 then Nothing else Just (n,n-1)) 10
  = <10,9,8,7,6,5,4,3,2,1>

unfoldrN :: Int -> (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector with at most n by repeatedly applying the generator function to the a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

 unfoldrN 3 (\n -> Just (n,n-1)) 10 = <10,9,8>

constructN :: Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements by repeatedly applying the generator function to the already constructed part of the vector.

 constructN 3 f = let a = f <> ; b = f <a> ; c = f <a,b> in f <a,b,c>

constructrN :: Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements from right to left by repeatedly applying the generator function to the already constructed part of the vector.

 constructrN 3 f = let a = f <> ; b = f<a> ; c = f <b,a> in f <c,b,a>

Enumeration

enumFromN :: Num a => a -> Int -> Vector a

O(n) Yield a vector of the given length containing the values x, x+1 etc. This operation is usually more efficient than enumFromTo.

 enumFromN 5 3 = <5,6,7>

enumFromStepN :: Num a => a -> a -> Int -> Vector a

O(n) Yield a vector of the given length containing the values x, x+y, x+y+y etc. This operations is usually more efficient than enumFromThenTo.

 enumFromStepN 1 0.1 5 = <1,1.1,1.2,1.3,1.4>

enumFromTo :: Enum a => a -> a -> Vector a

O(n) Enumerate values from x to y.

WARNING: This operation can be very inefficient. If at all possible, use enumFromN instead.

enumFromThenTo :: Enum a => a -> a -> a -> Vector a

O(n) Enumerate values from x to y with a specific step z.

WARNING: This operation can be very inefficient. If at all possible, use enumFromStepN instead.

Concatenation

cons :: Mixed u v a => a -> v a -> Vector a

O(n) Prepend an element

snoc :: Mixed u v a => v a -> a -> Vector a

O(n) Append an element

(++) :: (Mixed u v a, Mixed u' v' a) => v a -> v' a -> Vector a

O(m+n) Concatenate two vectors

concat :: Mixed u v a => [v a] -> Vector a

O(n) Concatenate all vectors in the list

Restricting memory usage

force :: Mixed u v a => v a -> Vector a

O(n) Yield the argument but force it not to retain any extra memory, possibly by copying it.

This is especially useful when dealing with slices. For example:

 force (slice 0 2 <huge vector>)

Here, the slice retains a reference to the huge vector. Forcing it creates a copy of just the elements that belong to the slice and allows the huge vector to be garbage collected.

Modifying vectors

Bulk updates

(//)

Arguments

:: Mixed u v a

=> v a
initial vector (of length m)

-> [(Int, a)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,a) from the list, replace the vector element at position i by a.

 <5,9,2,7> // [(2,1),(0,3),(2,8)] = <3,9,8,7>

update

Arguments

:: (Mixed u v a, Vector v' (Int, a))

=> v a
initial vector (of length m)

-> v' (Int, a)
vector of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,a) from the vector of index/value pairs, replace the vector element at position i by a.

 update <5,9,2,7> <(2,1),(0,3),(2,8)> = <3,9,8,7>

update_

Arguments

:: (Mixed u v a, Vector v' Int, Vector v'' a)

=> v a
initial vector (of length m)

-> v' Int
index vector (of length n1)

-> v'' a
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value a from the value vector, replace the element of the initial vector at position i by a.

 update_ <5,9,2,7>  <2,0,2> <1,3,8> = <3,9,8,7>

The function update provides the same functionality and is usually more convenient.

 update_ xs is ys = update xs (zip is ys)

unsafeUpd :: Mixed u v a => v a -> [(Int, a)] -> Vector a

Same as (//) but without bounds checking.

unsafeUpdate :: (Mixed u v a, Vector v' (Int, a)) => v a -> v' (Int, a) -> Vector a

Same as update but without bounds checking.

unsafeUpdate_ :: (Mixed u v a, Vector v' Int, Vector v'' a) => v a -> v' Int -> v'' a -> Vector a

Same as update_ but without bounds checking.

Accumulations

accum

Arguments

:: Mixed u v a

=> (a -> b -> a)
accumulating function f

-> v a
initial vector (of length m)

-> [(Int, b)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,b) from the list, replace the vector element a at position i by f a b.

 accum (+) <5,9,2> [(2,4),(1,6),(0,3),(1,7)] = <5+3, 9+6+7, 2+4>

accumulate

Arguments

:: (Mixed u v a, Vector v' (Int, b))

=> (a -> b -> a)
accumulating function f

-> v a
initial vector (of length m)

-> v' (Int, b)
vector of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,b) from the vector of pairs, replace the vector element a at position i by f a b.

 accumulate (+) <5,9,2> <(2,4),(1,6),(0,3),(1,7)> = <5+3, 9+6+7, 2+4>

accumulate_

Arguments

:: (Mixed u v a, Vector v' Int, Vector v'' b)

=> (a -> b -> a)
accumulating function f

-> v a
initial vector (of length m)

-> v' Int
index vector (of length n1)

-> v'' b
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value b from the the value vector, replace the element of the initial vector at position i by f a b.

 accumulate_ (+) <5,9,2> <2,1,0,1> <4,6,3,7> = <5+3, 9+6+7, 2+4>

The function accumulate provides the same functionality and is usually more convenient.

 accumulate_ f as is bs = accumulate f as (zip is bs)

unsafeAccum :: Mixed u v a => (a -> b -> a) -> v a -> [(Int, b)] -> Vector a

Same as accum but without bounds checking.

unsafeAccumulate :: (Mixed u v a, Vector v' (Int, b)) => (a -> b -> a) -> v a -> v' (Int, b) -> Vector a

Same as accumulate but without bounds checking.

unsafeAccumulate_ :: (Mixed u v a, Vector v' Int, Vector v'' b) => (a -> b -> a) -> v a -> v' Int -> v'' b -> Vector a

Same as accumulate_ but without bounds checking.

Permutations

reverse :: Mixed u v a => v a -> Vector a

O(n) Reverse a vector

backpermute :: (Mixed u v a, Vector v' Int) => v a -> v' Int -> Vector a

O(n) Yield the vector obtained by replacing each element i of the index vector by xs!i. This is equivalent to map (xs!) is but is often much more efficient.

 backpermute <a,b,c,d> <0,3,2,3,1,0> = <a,d,c,d,b,a>

unsafeBackpermute :: (Mixed u v a, Vector v' Int) => v a -> v' Int -> Vector a

Same as backpermute but without bounds checking.

Safe destructive updates

modify :: Mixed u v a => (forall s. u s a -> ST s ()) -> v a -> Vector a

Apply a destructive operation to a vector. The operation will be performed in place if it is safe to do so and will modify a copy of the vector otherwise.

 modify (\v -> write v 0 'x') (replicate 3 'a') = <'x','a','a'>

Elementwise operations

Indexing

indexed :: (Vector v a, Mixed u v (Int, a)) => v a -> Vector (Int, a)

O(n) Pair each element in a vector with its index

Mapping

map :: Vector v a => (a -> b) -> v a -> Vector b

O(n) Map a function over a vector

imap :: Vector v a => (Int -> a -> b) -> v a -> Vector b

O(n) Apply a function to every element of a vector and its index

concatMap :: (Mixed u v b, Vector v' a) => (a -> v b) -> v' a -> Vector b

Map a function over a vector and concatenate the results.

Monadic mapping

mapM :: (Monad m, Vector v a) => (a -> m b) -> v a -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results

mapM_ :: (Monad m, Vector v a) => (a -> m b) -> v a -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results

forM :: (Monad m, Vector v a) => v a -> (a -> m b) -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results. Equvalent to flip mapM.

forM_ :: (Monad m, Vector v a) => v a -> (a -> m b) -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results. Equivalent to flip mapM_.

Zipping

zipWith :: (Vector va a, Vector vb b) => (a -> b -> c) -> va a -> vb b -> Vector c

O(min(m,n)) Zip two vectors with the given function.

zipWith3 :: (Vector va a, Vector vb b, Vector vc c) => (a -> b -> c -> d) -> va a -> vb b -> vc c -> Vector d

Zip three vectors with the given function.

zipWith4 :: (Vector va a, Vector vb b, Vector vc c, Vector vd d) => (a -> b -> c -> d -> e) -> va a -> vb b -> vc c -> vd d -> Vector e

zipWith5 :: (Vector va a, Vector vb b, Vector vc c, Vector vd d, Vector ve e) => (a -> b -> c -> d -> e -> f) -> va a -> vb b -> vc c -> vd d -> ve e -> Vector f

zipWith6 :: (Vector va a, Vector vb b, Vector vc c, Vector vd d, Vector ve e, Vector vf f) => (a -> b -> c -> d -> e -> f -> g) -> va a -> vb b -> vc c -> vd d -> ve e -> vf f -> Vector g

izipWith :: (Vector va a, Vector vb b) => (Int -> a -> b -> c) -> va a -> vb b -> Vector c

O(min(m,n)) Zip two vectors with a function that also takes the elements' indices.

izipWith3 :: (Vector va a, Vector vb b, Vector vc c) => (Int -> a -> b -> c -> d) -> va a -> vb b -> vc c -> Vector d

Zip three vectors and their indices with the given function.

izipWith4 :: (Vector va a, Vector vb b, Vector vc c, Vector vd d) => (Int -> a -> b -> c -> d -> e) -> va a -> vb b -> vc c -> vd d -> Vector e

izipWith5 :: (Vector va a, Vector vb b, Vector vc c, Vector vd d, Vector ve e) => (Int -> a -> b -> c -> d -> e -> f) -> va a -> vb b -> vc c -> vd d -> ve e -> Vector f

izipWith6 :: (Vector va a, Vector vb b, Vector vc c, Vector vd d, Vector ve e, Vector vf f) => (Int -> a -> b -> c -> d -> e -> f -> g) -> va a -> vb b -> vc c -> vd d -> ve e -> vf f -> Vector g

zip :: (Vector va a, Vector vb b) => va a -> vb b -> Vector (a, b)

Elementwise pairing of array elements.

zip3 :: (Vector va a, Vector vb b, Vector vc c) => va a -> vb b -> vc c -> Vector (a, b, c)

zip together three vectors into a vector of triples

zip4 :: (Vector va a, Vector vb b, Vector vc c, Vector vd d) => va a -> vb b -> vc c -> vd d -> Vector (a, b, c, d)

zip5 :: (Vector va a, Vector vb b, Vector vc c, Vector vd d, Vector ve e) => va a -> vb b -> vc c -> vd d -> ve e -> Vector (a, b, c, d, e)

zip6 :: (Vector va a, Vector vb b, Vector vc c, Vector vd d, Vector ve e, Vector vf f) => va a -> vb b -> vc c -> vd d -> ve e -> vf f -> Vector (a, b, c, d, e, f)

Monadic zipping

zipWithM :: (Monad m, Vector va a, Vector vb b) => (a -> b -> m c) -> va a -> vb b -> m (Vector c)

O(min(m,n)) Zip the two vectors with the monadic action and yield a vector of results

zipWithM_ :: (Monad m, Vector va a, Vector vb b) => (a -> b -> m c) -> va a -> vb b -> m ()

O(min(m,n)) Zip the two vectors with the monadic action and ignore the results

Unzipping

unzip :: Vector v (a, b) => v (a, b) -> (Vector a, Vector b)

O(min(m,n)) Unzip a vector of pairs.

unzip3 :: Vector v (a, b, c) => v (a, b, c) -> (Vector a, Vector b, Vector c)

unzip4 :: Vector v (a, b, c, d) => v (a, b, c, d) -> (Vector a, Vector b, Vector c, Vector d)

unzip5 :: Vector v (a, b, c, d, e) => v (a, b, c, d, e) -> (Vector a, Vector b, Vector c, Vector d, Vector e)

unzip6 :: Vector v (a, b, c, d, e, f) => v (a, b, c, d, e, f) -> (Vector a, Vector b, Vector c, Vector d, Vector e, Vector f)

Working with predicates

Filtering

filter :: Mixed u v a => (a -> Bool) -> v a -> Vector a

O(n) Drop elements that do not satisfy the predicate

ifilter :: Mixed u v a => (Int -> a -> Bool) -> v a -> Vector a

O(n) Drop elements that do not satisfy the predicate which is applied to values and their indices

filterM :: (Monad m, Mixed u v a) => (a -> m Bool) -> v a -> m (Vector a)

O(n) Drop elements that do not satisfy the monadic predicate

takeWhile :: Mixed u v a => (a -> Bool) -> v a -> Vector a

O(n) Yield the longest prefix of elements satisfying the predicate without copying.

dropWhile :: Mixed u v a => (a -> Bool) -> v a -> Vector a

O(n) Drop the longest prefix of elements that satisfy the predicate without copying.

Partitioning

partition :: Mixed u v a => (a -> Bool) -> v a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The relative order of the elements is preserved at the cost of a sometimes reduced performance compared to unstablePartition.

unstablePartition :: Mixed u v a => (a -> Bool) -> v a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The order of the elements is not preserved but the operation is often faster than partition.

span :: Mixed u v a => (a -> Bool) -> v a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that satisfy the predicate and the rest without copying.

break :: (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that do not satisfy the predicate and the rest without copying.

Searching

elem :: (Vector v a, Eq a) => a -> v a -> Bool

O(n) Check if the vector contains an element

notElem :: (Vector v a, Eq a) => a -> Vector a -> Bool

O(n) Check if the vector does not contain an element (inverse of elem)

find :: Vector v a => (a -> Bool) -> v a -> Maybe a

O(n) Yield Just the first element matching the predicate or Nothing if no such element exists.

findIndex :: Vector v a => (a -> Bool) -> v a -> Maybe Int

O(n) Yield Just the index of the first element matching the predicate or Nothing if no such element exists.

findIndices :: Vector v a => (a -> Bool) -> v a -> Vector Int

O(n) Yield the indices of elements satisfying the predicate in ascending order.

elemIndex :: (Vector v a, Eq a) => a -> v a -> Maybe Int

O(n) Yield Just the index of the first occurence of the given element or Nothing if the vector does not contain the element. This is a specialised version of findIndex.

elemIndices :: (Vector v a, Eq a) => a -> v a -> Vector Int

O(n) Yield the indices of all occurences of the given element in ascending order. This is a specialised version of findIndices.

Folding

foldl :: Vector v b => (a -> b -> a) -> a -> v b -> a

O(n) Left fold

foldl1 :: Vector v a => (a -> a -> a) -> v a -> a

O(n) Left fold on non-empty vectors

foldl' :: Vector v b => (a -> b -> a) -> a -> v b -> a

O(n) Left fold with strict accumulator

foldl1' :: Vector v a => (a -> a -> a) -> v a -> a

O(n) Left fold on non-empty vectors with strict accumulator

foldr :: Vector v a => (a -> b -> b) -> b -> v a -> b

O(n) Right fold

foldr1 :: Vector v a => (a -> a -> a) -> v a -> a

O(n) Right fold on non-empty vectors

foldr' :: Vector v a => (a -> b -> b) -> b -> v a -> b

O(n) Right fold with a strict accumulator

foldr1' :: Vector v a => (a -> a -> a) -> v a -> a

O(n) Right fold on non-empty vectors with strict accumulator

ifoldl :: Vector v b => (a -> Int -> b -> a) -> a -> v b -> a

O(n) Left fold (function applied to each element and its index)

ifoldl' :: Vector v b => (a -> Int -> b -> a) -> a -> v b -> a

O(n) Left fold with strict accumulator (function applied to each element and its index)

ifoldr :: Vector v a => (Int -> a -> b -> b) -> b -> v a -> b

O(n) Right fold (function applied to each element and its index)

ifoldr' :: Vector v a => (Int -> a -> b -> b) -> b -> v a -> b

O(n) Right fold with strict accumulator (function applied to each element and its index)

Specialised folds

all :: Vector v a => (a -> Bool) -> v a -> Bool

O(n) Check if all elements satisfy the predicate.

any :: Vector v a => (a -> Bool) -> v a -> Bool

O(n) Check if any element satisfies the predicate.

and :: Vector v Bool => v Bool -> Bool

O(n) Check if all elements are True

or :: Vector v Bool => v Bool -> Bool

O(n) Check if any element is True

sum :: (Vector v a, Num a) => v a -> a

O(n) Compute the sum of the elements

product :: (Vector v a, Num a) => v a -> a

O(n) Compute the produce of the elements

maximum :: (Vector v a, Ord a) => v a -> a

O(n) Yield the maximum element of the vector. The vector may not be empty.

maximumBy :: Vector v a => (a -> a -> Ordering) -> v a -> a

O(n) Yield the maximum element of the vector according to the given comparison function. The vector may not be empty.

minimum :: (Vector v a, Ord a) => v a -> a

O(n) Yield the minimum element of the vector. The vector may not be empty.

minimumBy :: Vector v a => (a -> a -> Ordering) -> v a -> a

O(n) Yield the minimum element of the vector according to the given comparison function. The vector may not be empty.

minIndex :: (Vector v a, Ord a) => v a -> Int

O(n) Yield the index of the minimum element of the vector. The vector may not be empty.

minIndexBy :: Vector v a => (a -> a -> Ordering) -> v a -> Int

O(n) Yield the index of the minimum element of the vector according to the given comparison function. The vector may not be empty.

maxIndex :: (Vector v a, Ord a) => v a -> Int

O(n) Yield the index of the maximum element of the vector. The vector may not be empty.

maxIndexBy :: Vector v a => (a -> a -> Ordering) -> v a -> Int

O(n) Yield the index of the maximum element of the vector according to the given comparison function. The vector may not be empty.

Monadic folds

foldM :: (Vector v b, Monad m) => (a -> b -> m a) -> a -> v b -> m a

O(n) Monadic fold

foldM' :: (Vector v b, Monad m) => (a -> b -> m a) -> a -> v b -> m a

O(n) Monadic fold with strict accumulator

fold1M :: (Vector v a, Monad m) => (a -> a -> m a) -> v a -> m a

O(n) Monadic fold over non-empty vectors

fold1M' :: (Vector v a, Monad m) => (a -> a -> m a) -> v a -> m a

O(n) Monadic fold over non-empty vectors with strict accumulator

foldM_ :: (Vector v b, Monad m) => (a -> b -> m a) -> a -> v b -> m ()

O(n) Monadic fold that discards the result

foldM'_ :: (Vector v b, Monad m) => (a -> b -> m a) -> a -> v b -> m ()

O(n) Monadic fold with strict accumulator that discards the result

fold1M_ :: (Vector v a, Monad m) => (a -> a -> m a) -> v a -> m ()

O(n) Monadic fold over non-empty vectors that discards the result

fold1M'_ :: (Vector v a, Monad m) => (a -> a -> m a) -> Vector a -> m ()

O(n) Monadic fold over non-empty vectors with strict accumulator that discards the result

Monadic sequencing

sequence :: (Mixed u v (m a), Monad m) => v (m a) -> m (Vector a)

Evaluate each action and collect the results

sequence_ :: (Vector v (m a), Monad m) => v (m a) -> m ()

Evaluate each action and discard the results

Prefix sums (scans)

prescanl :: Vector v b => (a -> b -> a) -> a -> v b -> Vector a

O(n) Prescan

 prescanl f z = init . scanl f z

Example: prescanl (+) 0 <1,2,3,4> = <0,1,3,6>

prescanl' :: Vector v b => (a -> b -> a) -> a -> v b -> Vector a

O(n) Prescan with strict accumulator

postscanl :: Vector v b => (a -> b -> a) -> a -> v b -> Vector a

O(n) Scan

 postscanl f z = tail . scanl f z

Example: postscanl (+) 0 <1,2,3,4> = <1,3,6,10>

postscanl' :: Vector v b => (a -> b -> a) -> a -> v b -> Vector a

O(n) Scan with strict accumulator

scanl :: Vector v b => (a -> b -> a) -> a -> v b -> Vector a

O(n) Haskell-style scan

 scanl f z <x1,...,xn> = <y1,...,y(n+1)>
   where y1 = z
         yi = f y(i-1) x(i-1)

Example: scanl (+) 0 <1,2,3,4> = <0,1,3,6,10>

scanl' :: Vector v b => (a -> b -> a) -> a -> v b -> Vector a

O(n) Haskell-style scan with strict accumulator

scanl1 :: Mixed u v a => (a -> a -> a) -> v a -> Vector a

O(n) Scan over a non-empty vector

 scanl f <x1,...,xn> = <y1,...,yn>
   where y1 = x1
         yi = f y(i-1) xi

scanl1' :: Mixed u v a => (a -> a -> a) -> v a -> Vector a

O(n) Scan over a non-empty vector with a strict accumulator

prescanr :: Vector v a => (a -> b -> b) -> b -> v a -> Vector b

O(n) Right-to-left prescan

 prescanr f z = reverse . prescanl (flip f) z . reverse

prescanr' :: Vector v a => (a -> b -> b) -> b -> v a -> Vector b

O(n) Right-to-left prescan with strict accumulator

postscanr :: Vector v a => (a -> b -> b) -> b -> v a -> Vector b

O(n) Right-to-left scan

postscanr' :: Vector v a => (a -> b -> b) -> b -> v a -> Vector b

O(n) Right-to-left scan with strict accumulator

scanr :: Vector v a => (a -> b -> b) -> b -> v a -> Vector b

O(n) Right-to-left Haskell-style scan

scanr' :: Vector v a => (a -> b -> b) -> b -> v a -> Vector b

O(n) Right-to-left Haskell-style scan with strict accumulator

scanr1 :: Mixed u v a => (a -> a -> a) -> v a -> Vector a

O(n) Right-to-left scan over a non-empty vector

scanr1' :: (a -> a -> a) -> Vector a -> Vector a

O(n) Right-to-left scan over a non-empty vector with a strict accumulator

Conversions

Lists

toList :: Vector v a => v a -> [a]

O(n) Convert a vector to a list

fromList :: [a] -> Vector a

O(n) Convert a list to a vector

fromListN :: Int -> [a] -> Vector a

O(n) Convert the first n elements of a list to a vector

 fromListN n xs = fromList (take n xs)

Other vector types

convert :: (Vector v a, Vector w a) => v a -> w a

Mutable vectors

freeze :: (PrimMonad m, Mixed u v a) => u (PrimState m) a -> m (Vector a)

O(n) Yield an immutable copy of the mutable vector.

thaw :: (PrimMonad m, Mixed u v a) => v a -> m (MVector (PrimState m) a)

O(n) Yield a mutable copy of the immutable vector.

copy :: (PrimMonad m, Mixed u v a, Mixed u' v' a) => u (PrimState m) a -> v' a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length.

unsafeFreeze :: (PrimMonad m, Mixed u v a) => u (PrimState m) a -> m (Vector a)

O(1) Unsafe convert a mutable vector to an immutable one without copying. The mutable vector may not be used after this operation.

unsafeThaw :: (PrimMonad m, Mixed u v a) => v a -> m (MVector (PrimState m) a)

O(1) Unsafely convert an immutable vector to a mutable one without copying. The immutable vector may not be used after this operation.

unsafeCopy :: (PrimMonad m, Mixed u v a, Mixed u' v' a) => u (PrimState m) a -> v' a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length. This is not checked.