Data.Vector.Mixed.Internal

Documentation

data MVector where

A MVector s a is mutable vector that could have any vector type underneath

Constructors

MV :: Mixed mv v a => !(mv s a) -> MVector s a

mboxed :: MVector s a -> MVector s a

munboxed :: Unbox a => MVector s a -> MVector s a

data Vector where

Constructors

V :: Mixed mv v a => !(v a) -> Vector a

boxed :: Vector a -> Vector a

unboxed :: Unbox a => Vector a -> Vector a

class (Typeable2 mv, Typeable1 v, mv ~ Mutable v, MVector mv a, Vector v a) => Mixed mv v a | mv -> v, v -> mv where

Vector doesn't provide a way to recover the type of the immutable vector from the mutable vector type

This would otherwise prevent us from finishing the implementation of basicUnsafeFreeze in Vector

This class captures the invariants necessary to hide the choice of vector type from the user in such a way that we can go from mutable vector to immutabl vector and back again.

Methods

mmix :: mv s a -> MVector s a

mix :: v a -> Vector a